requirements.txt
blinker==1.9.0
click==8.1.8
defusedxml==0.7.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
fpdf2==2.8.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
